+++ src/main.rs
@@ -1,5 +1,5 @@
 use clap::Parser;
-use realmenv::{Cli, CliHandler};
+use realm::{Cli, CliHandler};
 use std::process;
 
 #[tokio::main]